explicitly cast the codimension/dimension enums to int in order to find the correct specialization
[[Imported from SVN: r7599]]

+++ dune/grid/io/file/dgfparser/test/main.cc
@@ -131,7 +131,7 @@ try {
       }
     }
 
-    nofVtxParams = gridPtr.nofParameters( GridType::dimension );
+    nofVtxParams = gridPtr.nofParameters( (int) GridType::dimension );
     if( nofVtxParams > 0 )
     {
       std::cout << "Reading Vertex Parameters:" << std::endl;
